# Conflicts:
#   client/src/main/java/client/Main.java
#   client/src/main/java/client/scenes/MainCtrl.java
#   client/src/main/java/client/utils/ServerUtils.java

+++ server/src/main/java/server/service/BoardService.java
@@ -9,6 +9,7 @@ import server.database.BoardRepository;
 
 import javax.transaction.Transactional;
 import java.util.List;
+import java.util.Optional;
 
 
 @Service
@@ -42,18 +43,30 @@ public class BoardService {
      * @param board to delete
      */
     public void delete(Board board){
-        repo.delete(board);
-
+        if(repo.existsById(board.getId())){
+            repo.delete(board);
+        }
     }
 
     /**
      * return board instance given its id
+     *
      * @param id board id
      * @return board object
      */
     public Board getById(int id){
-        Board board = repo.findById(id).get();
-        return board;
+        return repo.getById(id);
+    }
+
+    /**
+     * return board instance given its id
+     *
+     * @param id board id
+     * @return board object
+     */
+    public Optional<Board> findById(int id){
+        return repo.findById(id);
+
     }
 
     /**
